+++ ipwhois/__init__.py
@@ -22,7 +22,7 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-__version__ = '0.8.3'
+__version__ = '0.9.0'
 
 from .ipwhois import (IPWhois, IPDefinedError, ASNLookupError,
                       ASNRegistryError, WhoisLookupError, HostLookupError)